Add BitmapHelpers
This commit introduces BitmapHelpers.
The module will be responsible to hold shared methods across the testing
suite.
