+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require 'pry'
+require 'pathname'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
@@ -10,9 +11,14 @@ RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
 
+  config.add_formatter 'documentation'
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
   #     --seed 1234
   config.order = 'random'
+
+  # Require all ruby helper files inside spec/support folder
+  Dir[Pathname.pwd.join('spec', 'support', '**', '*.rb')].each { |f| require f }
+  config.include BitmapHelpers
 end
